refactor(e2e): extract waitForInitialScan + add CI job timeout
- Add `waitForInitialScan(page, timeoutMs?)` to `e2e/helpers/redux.ts`
  and replace 11 inline `appWindow.waitForFunction(() => …)` blocks
  across copy / delete / bulk-delete / regression / unlink-agent specs.
  Net: 146 deletions / 60 insertions, -86 LOC. Behavior unchanged
  (same selector predicate, same 10s default timeout).
- Add `timeout-minutes: 30` to the e2e job in
  `.github/workflows/e2e.yml`. Caps a hung Playwright run at 30 min
  instead of consuming the runner's 6-hour default.
