+++ .github/workflows/e2e.yml
@@ -23,6 +23,7 @@ concurrency:
 jobs:
   e2e:
     runs-on: macos-latest
+    timeout-minutes: 30
 
     steps:
       - name: Checkout
